copy prop changes
